requirements.txt
flask
python-dotenv
openai-whisper
google-generativeai
ffmpeg-python
moviepy